Slash Command, CleanUp
